main stylesheet moved to /assets/style/index.js
