+++ screens/Login/Signup.js
@@ -18,7 +18,7 @@ import { Overlay, Card } from "react-native-elements";
 import * as Yup from 'yup';
 import { Ionicons } from '@expo/vector-icons';
 
-import { loginStyles } from '../../components/Styles';
+import { loginStyles } from '../../assets/style';
 import { isUpdateExpression } from "@babel/types";
 import { gql } from "apollo-boost";
 import { useApolloClient, useMutation, useQuery } from "@apollo/react-hooks";
