+++ screens/Login/SignupBpII.js
@@ -17,7 +17,7 @@ import { Overlay, Card, Button } from "react-native-elements";
 import * as Yup from 'yup';
 import { Ionicons } from '@expo/vector-icons';
 import { UserContext } from "../../context/userState";
-import { loginStyles } from '../../components/Styles';
+import { loginStyles } from '../../assets/style';
 import { isUpdateExpression } from "@babel/types";
 import { gql } from "apollo-boost";
 import { useApolloClient, useMutation, useQuery } from "@apollo/react-hooks";
